Add all the build stuff for saidar (assuming Pete sticks with that name). The configure script will attempt to find curses, but if it can't will simply disable building of saidar (after giving a warning).
Also made the building of saidar, statgrab, and the examples
conditional. They're enabled by default, but can be disabled using a
configure flag. Whilst I was at it, added a configure flag to disable
manpage installation at configure time.

I'm not sure how well the curses detection works; we've had at least
one report of a problem with it. Testing will be required.