Basic implementation of name mapping for solaris. The code is ugly, mostly because its initially taken from a sun example, and modified. It needs tiding somewhat. I'm not sure if tools.c is the right place for this, i guess it could go in disk_stats.c, but its doing a specific job, and we could right another mini-program which does the disk mappings.
Comments welcome about the general implementation.

Should we call build_mapping if the user calls get_svr_from_bsd and
build_mapping hasn't been run before? Or should we force them to use the
init function? (Bare in mind, on some systems, to do the mappings correctly
the program will require to be root. (Only ones with a /dev/osa , so everything
with an A1000 for e.g.)