This has been a bit more complex than it probably should have been. We now have the command line arguments on FreeBSD5, NetBSD, and probably OpenBSD without using kvm. The sysctl interface provided the required information.
However, on FreeBSD 4 and Dragonfly BSD there was no matching sysctl
interface, so the kvm way remains. Fortunately neither of those two OS's
require elevated privileges to use that bit of kvm.

Finally in a decent position with no elevated privileges required for this
section of code, on any of the BSD's.

Still one bug to fix: there are 5 "null" processes on the end of the list.
This has been introduced at some point, because it wasn't there yesterday.