Fix the strict-aliasing warnings.
This is even uglier than the previous code, since it pushes a test for a
sentinel value into the VECTOR_RESIZE macro. I'm pretty certain that the
pointer-to-static-used-as-sentinel is my own invention. I'm not entirely
convinced that it's a good idea, but I can't see anything technically
wrong with it...