Only enable setuid/setgid binaries if requested.
Prior to this commit, if required to collect statistics by the
operating system, binaries were installed with the setuid or setgid
bit set. Since the actual cases where this is needed are rare, and
since it leaves a hole for exploitation, we now only do this if the
user explicitly requests it when building.

So setuid and setgid binaries are still available, you just need to
use --enable-setuid-binaries or --enable-setgid-binaries to enable
them. These options have no effect if the operating system has no need
for setuid or setgid binaries to collect statistics.