Move the ChangeLog generation out of the Makefile.
Having it in the Makefile broke the distcheck target. The specific
failure case is if you do a VPATH configure, and then use the make
dist target, it regenerates the ChangeLog file rather than using the
one in $(srcdir) as it does for other distributed files. The reason it
does this is because there's a ChangeLog target in the Makefile.

This doesn't feel like a perfect solution, but it only makes life a
little harder for devs, and fixes potential breakages for users.