Display new device_canonical field.
A new field has been added to the filesystem statistics in libstatgrab.
The device_canonical field holds the real device name after resolving
symlinks. The original device_name field remains unchanged and holds
whatever is given by the operating system. When device_name is not a
symlink both fields are the same.

Here's an example of where it's useful:

/dev/disk/by-uuid/ee2f048d-68f6-428a-a95a-c949d3c05d67 -> ../../sda2

In this case, device_name contains the left-hand side and
device_canonical contains "/dev/sda2".