Fix tests when using a VPATH build.
The problem is that App::Prove (or whatever it uses) doesn't chdir
to the directory before running the tests. So the relative include
path in the .t file is only valid if you're inside the single_threaded
or multi_threaded directories. Running the tests from there works
fine, but if you run it from the directory above the relative path
points to the wrong place and it breaks. This change uses an absolute
path so it works in both cases.

An alternative fix would be to modify the Makefile in tests to chdir
to each sub-directory before launching App::Prove, or if App::Prove
has a way of doing the chdir itself that'd also work.

Fixes #43.