Solaris: don't skip interfaces with partial data.
In testing of VLANed interfaces it became apparent that they don't have
all data items (notably "collisions") which resulted in us skipping over
the interface. I've reworked this to only include the data items that
are present and not skip the interface.

This also fixes a bug caused by us increasing the vector size and
partially populating it before skipping on to the next interface. Because
the interfaces counter wasn't incremented the next loop around would
overwrite this partial entry. This caused at best confusion if the skipped
interfaces were last in the list; "why is only the last of these odd
interfaces appearing?", and at worst an interface entry with a mixture
of a skipped interface and an actual one. By not skipping interfaces we
avoid this problem.