Allow version to be overridden.
By default, it still uses the short version of the commit reference. This
is most useful for normal CI builds so you can easily see which commit
a tarball is made from. But when testing for a release one might want
to specify the version explicitly, so this can now be overridden by
manually triggering a build and setting LSG_VERSION (eg. 0.92).