Fix build/install of manpages.
This defaults to attempting to install manpages, unless --disable-man
is given to stop it. It defaults to not building the manpages, unless
it finds docbook2man to build them with.

So for users of the release tarball this will install the manual pages
from the tarball, unless requested not to.

For users of the git repository it will error if they don't have
docbook2man, unless they choose to disable manual pages. I think this is
reasonable because docbook2man is a required tool for build from source.